pyproject.toml
[project]
name = "podwise"
version = "0.1.0"
description = "Agentic RAG pipeline for YouTube/podcast transcripts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-anthropic>=0.2",
    "langchain-community>=0.3",
    "langchain-chroma>=0.2",
    "langchain-experimental>=0.3",
    "langchain-ollama>=0.2",
    "chromadb>=0.5",
    "youtube-transcript-api>=0.6",
    "python-dotenv>=1.0",
    "typer[rich]>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
]
